+++ src/list.js
@@ -31,12 +31,20 @@
 
 function list(T, name) {
 
+  assert(isType(T), 'bad type');
+
   name = name || format('list(%s)', getName(T));
 
   function List(value, mut) {
 
     forbidNewOperator(this, List);
+
     assert(Arr.is(value), 'bad %s', name);
+
+    // makes List idempotent
+    if (value.every(T.is)) {
+      return value;
+    }
 
     var arr = [];
     for (var i = 0, len = value.length ; i < len ; i++ ) {